Remove flakiness-causing static initializer test
Having both of these tests makes them order dependent. If the one that sets the
final field is run first, the one checking for null will clearly fail (nothing
is resetting this static state). Unless you run a separate JVM per test method,
this will always be flaky (it's dependent on the `getDeclaredMethods` method in
the JVM and that returns a nondeterministic order).

Besides, the second test already covers the condition from the first.