Fix resource deferring to decouple tests from Surefire and Maven
`MockClassLoaderTest` runs fine with Maven but fails with buck.
The failure has to do with how it constructs a `URLClassLoader`.

With Maven, the current ClassLoader is an `AppPathClassLoader` (since Maven
runs the tests from an exploded folder tree) while with buck, it's a
`JarClassLoader`. You can't specify a `URLClassLaoder` to a non-root location
in a `jar`, it gets silently truncated to the root of the `jar`.

Therefore, introduce `ResourcePrefixClassLoader` which prefixes resource paths
(shocking, I know). This maintains the invariant that the test should fail to
find the resources if `MockClassLoader` is not deferring correctly.