Fix #1034 PowerMockRunner with mock-maker-inline Classloader issues
The root cause of the error is still in the class loader - org.powermock.core.classloader.MockClassLoader:

When executing a test annotated with RunWith(PowerMockRunner.class) where mockito-inline and powermock-api-mockito2 dependencies coexists, the invoking of MockClassLoader.getResources with "mockito-extensions/org.mockito.plugins.MockMaker" as argument returns an instance of Enumeration<URL> whose first element is the URL of mockito-inline. So the class loader will load the unexpected InlineByteBuddyMockMaker, which is defined in the mockito-inline jar. Whereas when executing a test annotated with RunWith(PowerMockRunner.class), the needed MockMaker should be based on the mockito-extensions/org.mockito.plugins.MockMaker file from the powermock-api-mockito2 jar: https://github.com/powermock/powermock/blob/release/2.x/powermock-api/powermock-api-mockito2/src/main/resources/mockito-extensions/org.mockito.plugins.MockMaker

As per the implementation, based on the document of ClassLoader, the method to override should be the protected findResources in stead of the public getResources and, the current implementation is just an extra if condition. In future if more if condition is to be added, refactor will be needed