Introduce PowerMockExtension for JUnit Jupiter support
This commit adds a new module that includes a JUnit Jupiter extension
implementation that relies on the existing Java agent to transform
classes and implements the same logic as the corresponding
PowerMockRule. More test coverage should be added before releasing this
in earnest but it should work just like the JUnit 4 rule.